Tried to fix the navbar & cart icon issues but ruined it even more.
The Cart icon that I made dynamic appears right above the product list component instead of on the header; it is because I am passing the functionality directly there. When I tried to implement it in the App.jsx instead, the whole thing got ruined even more. That is an issue that should be fixed.